backend/pyproject.toml
[project]
name = "notepipe-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "supabase>=2.10.0",
  "httpx>=0.27.0",
  "anthropic>=0.40.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.6.0",
  "tenacity>=8.2.0",
  "pypdf>=4.0.0",
  "python-docx>=1.1.0",
  "python-multipart>=0.0.9",
  "slowapi>=0.1.9",
]

[dependency-groups]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pytest-httpx>=0.30.0",
  "ruff>=0.8.0",
  "mypy>=1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]   # pycodestyle, pyflakes, isort, pyupgrade

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
